Perbaikan Penarikan dan topup Jopay

+++ src/tubespbo/Controller/Controller.java
@@ -5,7 +5,6 @@ import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
-import java.text.DecimalFormat;
 import java.text.SimpleDateFormat;
 import java.time.LocalDate;
 import java.util.ArrayList;
@@ -731,7 +730,7 @@ public class Controller {
                 user.setUser_name(rs.getString("user_name"));
                 user.setUser_pass(rs.getString("user_pass"));
                 user.setUser_role(rs.getString("user_role"));
-                user.setUser_wallet(rs.getDouble("user_wallet"));
+                user.setUser_wallet(rs.getFloat("user_wallet"));
             }
             return user;
         } catch (SQLException e) {
